refactor(providers): share neutral event mapping
